Refactored into Bullet class
